OpenGL -> DX12

+++ Ngin/src/Ngin/Renderer/Framebuffer.h
@@ -13,7 +13,7 @@ namespace Ngin {
 		virtual void Unbind() const = 0;
 		virtual void Resize(uint32_t width, uint32_t height) = 0;
 
-		virtual uint32_t GetColorAttachmentID() const = 0;
+		virtual uint64_t GetColorAttachmentID() const = 0;
 		virtual uint32_t GetWidth()  const = 0;
 		virtual uint32_t GetHeight() const = 0;
